Update cloud-box aliases
Renaming cb-code to cb-groups considering that's the project it's opening. I'm also noticing that VSCode is opening too quickly, meaning the cloud-box is not totally set up and ready for a SSH connection by the time the editor connects, and I'm forced to click "Retry" each time. Adding a sleep command will give that little breathing room.