Restructure profiles installation
This updates the entire profile installation strategy around the idea that
global configuration lives at the top-level of the project, and profile-specific
configuration lives in the respective `profiles/NAME` folder.

For bin scripts, git config, and zsh config, this means the use of "include"
files and directories that the profiles must link.

This will make it easier in the future to know how to apply global changes
versus profile-specific ones.

I'm sure there is a way to extract a `global install.conf.yaml`, but I didn't
think it was worth it at this point. I can live with the duplication.