Add setup scripts to profiles
I wanted a way to automatically install any system requirements for
each profile. For now, I'm going with the convention of running a
profile setup.sh script manually.

I considered making it part of the install.conf.yaml, but it feels like
it needs to be a preliminary step because things like zsh/git/brew
need to be available for the dotbot installation to run correctly.