Modularize Neovim configuration
Now that I'm getting a better handle on how the Neovim configuration
comes together and the customizations I'd like to make, it's time to
split this out into individual files within a core/ folder for
high-level vim config, and a plugins/ folder to configure each lazyvim
plugin. This helps keep things isolated and easily findable.