Back to WezTerm from kitty
I was having issues in kitty with emojis rendering properly when
SSHedinto my cloud-box. I would see ??? ???? instead of a laptop emoji.
Thisissue caused me to tinker with WezTerm once more, and I think I have
itat a good spot now. I was having issues with how the font
renderinglooked before but now I like it. I'll stick with it for a
while.