Bring back custom modularized Neovim config
I experimented with the LazyVim project, and while it was an eye-opener
as to what's possible with a Neovim setup, it was a little too much for
me right now.

Not all of the plugins were necessary for me at the moment, and there
were a handful of shortcuts that didn't make the most sense to me and I
don't necessarily want to override all of them.

I think moving forward the LazyVim project will be a great source of
inspiration as to what I can bring into my own setup, but I'll keep
things minimal for now and bring new features in as I find them
valuable.