Update dotfiles git origin in setup script
I tried to make edits to my dotfiles in my cloud-box and push those
changes, but then it asked for HTTP authentication when I only use SSH
authentication. I figured this was due to using a HTTP URL for the git
origin. Changing it to an SSH address fixed it. I'm going to add a step
to my setup scripts to use an SSH address. The reason I'm not doing it
within in the clone step, is because I don't think I can assume SSH
authentication will have been set up and available at the time these
scripts are running. The repo is public so I know it'll be available to
clone on HTTP, but I really only need SSH for future changes so I'm
updating the URL after the fact.