Add Neovim LSP config for Ruby
I'd like to use ruby-lsp over solargraph, with rubocop for linting and
syntax_tree for formatting. I'm also opting out of Mason installing
these tools as that could lead to weirdness as to where Neovim looks for
these commands. I'd prefer it to reach for the project bundle first to
respect explicit versioning.