Fix alt key behavior with Wezterm
I was trying to use an Alt+c Zellij keybinding with the more comfortable
right Alt, but it was not behaving the same. It would input something
like <ffffff><fffff> into the terminal. Some googling led me to this
config which gives the right Alt the same default behavior as the left
Alt

https://wezfurlong.org/wezterm/config/lua/config/force_reverse_video_cursor.html