More lazygit tweaks
I noticed the use of plenary in Neovim caused differences in coloring.
In the terminal, lazygit borders and text were white, but in Neovim they
were yellow. I don't like the inconsistency so I'll forgo that option.

I expanded the sidebar a little bit to make more room for commit titles,
and reduced the window scaling a little bit in Neovim so it's clearer
it's in the context of Neovim.