Remove WezTerm term config
This config was causing issues with the zsh-vi-mode plugin where the
cursor was always a block rather than a beam on insert mode, and block
on normal mode.

I traced it down to an issue with the having the custom wezterm terminfo
(for some reason, I didn't care to dig much further).