Don't format JSON in Neovim
I was making a single line change, and it auto-formatted the file so all
of the key values lined up column wise, and this was just not what I
wanted to do. I wanted to make the single line change, and I don't even
necessarily agree with that formatting rule in theory.