Update ZSH aliases
I'm going with a `v` theme for Neovim commands

There are times I try to run `bat cache --build` which is necessary for
my custom bat theme, but then I forget the `bat` alias with option flags
is messing that command up. I'll stick with `cat` and keep `bat` behind
the scenes.