Rename nvim config folder to nvim-lazyvim
I'm going to experiment with a new Neovim config, but I want the ability
to switch back to my tride and true lazyvim config until I feel like a
new config can become the default.

I'll start by renaming the current config to nvim-lazyvim so it's clear
which is which.