Rework ZSH config
There was a company recommendation as to when certain tools should be
loaded in regards to PATH. This led me to research more about the use of
zshenv, zprofile, and zshrc, which led me to believe I don't need a
zshenv other than customizing the zsh config location. Everything
environment/PATH related will live in zprofile, and anything relating to
interactive shells will live in my zshrc.