Fix zshenv
There was a source line to a file that didn't exist anymore. I also
moved the zshenv to the zsh folder. The zshenv file must exist in the
home folder in order to customize the zsh dotfiles location, but it's
okay to be nested in with the rest of the zsh config files in my
dotfiles repo.