Override groups command for cd
Because I have auto cd set up for my ~/Code folder, I can just type the
name of the project I want to enter without the need for an explicit
`cd`; however this is not the case for my primary project `groups`
because there is a conflicting `groups` command from the coreutils
package. By setting an alias, I can force the `cd` action.