Only load gh-navigator with git repo
Whenever I open Neovim in a directory that doesn't have a git repo, I
get a warning notification stating the plugin is only for that scenario.
I can add this condition to the lazy plugin so it won't load without a
git repo.

Note: I was getting linting errors on using vim.loop. AI explained
vim.loop is not standard lua but exposed by Neovim's Lua API and
available at runtime, but static analyzers can't know that. This is safe
to disable this warning.