Stop using /etc ZSH files for login shells
I was noticing, not in interactive shells, but in login shells created
by Claude Code, that the system bin paths were being shoved to the top
instead of my homebrew or devbox paths. After some research, I realized
this was because Claude Code login shells automatically use
/etc/zprofile which uses a macOS path_helper that puts those system
paths to the front. I am disabling this option in zshenv so I can avoid
that behavior and lean on my package managers first and foremost.