Fix nvim mini.tabline auto-hide
I was noticing if closing all the buffers in succession, closing the
last real buffer would open a new blank buffer, but the tabline would
still be visible. I don't want the tabline unless there's 2+ buffers. I
updated the behavior to be scheduled so the timing of Neovim opening a
new buffer wouldn't affect the buffer count calculation. I also updated
the autocmd events list because I only care about this when
adding/removing buffers. It doesn't need to run when entering buffers.